pyproject.toml
[tool.poetry]
name = "lsmfapi"
version = "0.1.0"
description = "Lenticularis SwissMeteo Forecast API"
authors = ["Raphael Eymann <r+claude@eymanns.ch>"]
packages = [{include = "lsmfapi", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115"
uvicorn = {extras = ["standard"], version = "^0.30"}
pydantic = "^2.9"
httpx = "^0.27"
apscheduler = "^3.10"
sqlalchemy = "^2.0"
cfgrib = "^0.9.15"
xarray = "^2024.0"
eccodes = "^2.38"
eccodes-cosmo-resources-python = "^2.38"
scipy = "^1.14"
numpy = "^1.26"
pyyaml = "^6.0"
aiofiles = "^24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"
